Windows Installation
This is a guide I wrote for myself on how to install plutus playground on Windows. To be honest, I was having huge problems following and understanding the current guide named WindowsWSL. Through many iterations and frustration I landed with this guide, which doesnt use docker or plutus-starter. I have tested the guide on 4 seperate windows machines to find tips and adjustments when things don't work exactly as expected. 

The guide tries to keep installation to a bare minimum, but there are optional sections at the end to install a nice editor, terminal and docker as well. There are also specific instructions for pioneers on exactly how to check out the correct commit for the current week.

I hope it will be of use to some of the future pioneers or anyone else who wishes to run the playground in windows.